Add alternative lesson structure
This notebook takes a different approach to teaching the basics (suggested by Matt): starting with the very concrete problem of reading a file that needs to be processed in some way. This provides a worked example that can be followed through the entire notebook.

The structure is a little more haphazard, because it introduces new concepts and primitives/data structures as they are needed, rather than giving a syntax tour to start with. This also contains much more of a narrative thread, which is written into the notebook.

It may elide some important details in places, but should cover the same concepts as the other notebook, just in a very different order.  The breaks between sections still need to be determined, because this flows a lot more smoothly from one section to the next.