Create Mathematics_and_Strings.ipynb
Basic coverage of maths and strings. This should be done on the command line beforehand, but maybe we do strings in the notebook?

Anyway, this covers a couple of important things. We can expand this if needed, but it is the same content, with some additional scaffolding to let students go through it on their own.