Update Multidimensional_data.ipynb
Added notes to the end of the notebook, pointing at the docs for numpy and scipy. Might be worth pointing folks at pandas and xarray, but it might derail the course somewhat.

Ideally this needs a couple more exercises to practice things like indexing, not just the one towards the end.