Switch to `mosaic` for goal plot
Use `mosaic` to make the grid. This is much more accessible than the previous method, using gridspec. This should make it easier for students to follow the code, and might be more useful for other things that they might want to do on their own.

See here for details of `mosaic`: https://matplotlib.org/3.5.1/tutorials/provisional/mosaic.html