New control file
This is a draft, but is now using the new notebooks and structure. We will need to look at the timing still, in particular we might have a lot of time on the last day, but we have the other notebooks to pull in if we are by some miracle actually ahead of schedule.