Update the final plot notebook
Add an image of the final plot to the start of the notebook. We can also use this to show early in the course.

Have also changed the final plot to not be in a function, which is a bit nicer to read and probably more sensible. We can revisit this as a function later.