Add missing dependencies to geocomp.yaml
Add pooch into the environment. It is used in the final plot notebook, and is referenced in a fair few others, especially the IO intro. It is a useful thing to have available for getting more data in case we are asked to show something specific.

`gio` is used in the final plot's notebook, as well as being shown in the IO notebook.

Added the text versions of the Mt St Helens data back into the initial download, since we often use it on the console.