Fix mangling of Panoma data's depths
`STUART` is already in feet, so we can leave it alone.
The other wells are all in metres, so changing `SHANKLE` and `NEWBY` to metres does not seem to work. We can convert them to feet though.

This gives us four wells in feet: `CRAWFORD`, `STUART` (both read in in feet), `SHANKLE` and `NEWBY` (converted from metres).