Create Cleaning_panoma.ipynb
Mostly complete notebook to clean up the mangled Panoma dataset. There are a couple of questions that still need to be answered before this is 'complete', mostly around the `Facies` column. We could also make all the column names one word, but that is going to badly break the `pandas for ml` notebook which assumes `'Well Name'` in particular.

I will put some text referring to tidy data and such as well, just as links for students to read on their own, more than for us to go through in class.