Add additional notebooks to EDA curriculum
While these may not be the final version of notebooks, this at least covers the main topics listed. We will probably have to see how time goes as to whether we are able to spend much time on the notebooks listed as `extras`.
My feeling is that we will probably be able to do at least one of them, depending on the interests of a given class and their background familiarity with pandas, in addition to the core pandas EDA + feature engineering content.

TODO: Some of these notebooks still need to be moved from `dev` into `prod`.