Add Intro to pandas notebook to control files
This notebook is split off from the existing Pandas for ML data Management one, so any course that includes that notebook should probably have this 'new' one as well.

I have added them as `extras` for now, rather than making them an explicit part of the curriculum. This is easy to change, but at least we are not missing material now.