Update Intro_to_pandas.ipynb
Minor tweaks to the Intro to pandas notebook:
Merged some markdown cells, expanded a couple of explanations by a few words, added a broadcasting example to make a new column.

This still does not have exercises, but assuming that the next notebook is the Cleaning data with pandas one, there is plenty of code to write there, so that might be ok.