Change kernel for EDA notebooks
These should all use `geoml` now, not `geocomp`.

The only notebook that has _not_ been updated is `Read_and_process_images`, which is included as an extra in the `geocomp` course. It is probably better to keep the `geocomp` environment for that course, since it has more complete beginners taking it than the EDA course, so hopefully changing a kernel is more natural by the time people get to the EDA course.