Add exercise and Excel loading to Intro to Pandas
Loading a `xlsx` document is a nice change. We will still write out a CSV locally, but we can carry on with either, I think.

Also adds an example of why categoricals can be nice, even if it does not go into it in detail.

This might still need some extra exercises, perhaps on the loaded data?