Update Cleaning_data_with_Pandas.ipynb
Start by loading a `xlsx` file.

Cleaned up the processing for Depth and Marine columns to be much nicer (no function, just using `.loc` better).

Not dropping the `RelPos` column any more, we can do it at the end of the Feature Engineering notebook.

We will now use this dataset, as written out in this notebook, in the Feature Engineering notebook going forward.