Update Create_messy_Panoma_data.py
Save a `.xlsx` version of the dataset. The first worksheet contains some metadata, so if you want to read the data, you will need `sheet_name='data'` as an argument to `read_excel`.

This also leaves the `Facies` as a series of numbers. They were being changed and then changed back. They are now left as-is.